+++ FireAlt.Mosaic.Authoring/Authorings/TilemapAuthoring.cs
@@ -8,6 +8,9 @@ namespace FireAlt.Mosaic.Authoring
     public class TilemapAuthoring : MonoBehaviour
     {
         public IntGridDefinition intGrid;
+        [Tooltip("True: to access the runtime IntGrid of this tilemap, the IntGridDefinition can be used directly.\n\n" +
+                 "False: the hash of the runtime IntGrid is unique per each instance and the entity needs to be queried to get the runtime hash.")]
+        public bool isGlobal = true;
         public RenderingData renderingData;
 
         public class Baker : Baker<TilemapAuthoring>
@@ -27,8 +30,9 @@ namespace FireAlt.Mosaic.Authoring
                 var refSprite = new RefSprite();
                 
                 BakerUtils.AddTilemapTransform(this, entity, authoring.renderingData, gridAuthoring);
-                BakerUtils.AddIntGridLayerData(this, entity, authoring.intGrid, refSprite, false, ref tilePivot, ref tileSize);
-                BakerUtils.AddRenderingData(this, authoring.gameObject, entity, authoring.intGrid.Hash, authoring.renderingData, refSprite);
+                BakerUtils.AddIntGridLayerData(this, entity, authoring.intGrid, authoring.isGlobal, refSprite,
+                    false, ref tilePivot, ref tileSize, out var runtimeHash);
+                BakerUtils.AddRenderingData(this, authoring.gameObject, entity, runtimeHash, authoring.renderingData, refSprite);
             }
         }
     }